use shared Db fixture for participants and match databases
