+++ match/tests/Piipan.Match.Core.IntegrationTests/MatchRecordDaoTests.cs
@@ -26,7 +26,7 @@ namespace Piipan.Match.Core.IntegrationTests
                 .ReturnsAsync(() =>
                 {
                     var conn = Factory.CreateConnection();
-                    conn.ConnectionString = ConnectionString;
+                    conn.ConnectionString = CollabConnectionString;
                     conn.Open();
                     return conn;
                 });
@@ -40,7 +40,7 @@ namespace Piipan.Match.Core.IntegrationTests
             using (var conn = Factory.CreateConnection())
             {
                 // Arrange
-                conn.ConnectionString = ConnectionString;
+                conn.ConnectionString = CollabConnectionString;
                 conn.Open();
                 ClearMatchRecords();
 
@@ -71,7 +71,7 @@ namespace Piipan.Match.Core.IntegrationTests
             using (var conn = Factory.CreateConnection())
             {
                 // Arrange
-                conn.ConnectionString = ConnectionString;
+                conn.ConnectionString = CollabConnectionString;
                 conn.Open();
                 ClearMatchRecords();
 
@@ -104,7 +104,7 @@ namespace Piipan.Match.Core.IntegrationTests
             using (var conn = Factory.CreateConnection())
             {
                 // Arrange
-                conn.ConnectionString = ConnectionString;
+                conn.ConnectionString = CollabConnectionString;
                 conn.Open();
                 ClearMatchRecords();
 
@@ -134,7 +134,7 @@ namespace Piipan.Match.Core.IntegrationTests
             using (var conn = Factory.CreateConnection())
             {
                 // Arrange
-                conn.ConnectionString = ConnectionString;
+                conn.ConnectionString = CollabConnectionString;
                 conn.Open();
                 ClearMatchRecords();
 
@@ -165,7 +165,7 @@ namespace Piipan.Match.Core.IntegrationTests
             using (var conn = Factory.CreateConnection())
             {
                 // Arrange
-                conn.ConnectionString = ConnectionString;
+                conn.ConnectionString = CollabConnectionString;
                 conn.Open();
 
                 var logger = Mock.Of<ILogger<MatchRecordDao>>();
@@ -209,7 +209,7 @@ namespace Piipan.Match.Core.IntegrationTests
             using (var conn = Factory.CreateConnection())
             {
                 // Arrange
-                conn.ConnectionString = ConnectionString;
+                conn.ConnectionString = CollabConnectionString;
                 conn.Open();
 
                 var logger = Mock.Of<ILogger<MatchRecordDao>>();
@@ -255,7 +255,7 @@ namespace Piipan.Match.Core.IntegrationTests
             using (var conn = Factory.CreateConnection())
             {
                 // Arrange
-                conn.ConnectionString = ConnectionString;
+                conn.ConnectionString = CollabConnectionString;
                 conn.Open();
 
                 var logger = Mock.Of<ILogger<MatchRecordDao>>();
