+++ match/tests/Piipan.Match.Func.Api.IntegrationTests/DbFixture.cs
@@ -4,6 +4,7 @@ using Dapper;
 using Npgsql;
 using Piipan.Match.Core.Models;
 using Piipan.Participants.Core.Models;
+using Piipan.Shared.Database;
 
 namespace Piipan.Match.Func.Api.IntegrationTests
 {
@@ -12,136 +13,8 @@ namespace Piipan.Match.Func.Api.IntegrationTests
     /// Creates a fresh set of participants and uploads tables, dropping them
     /// when testing is complete.
     /// </summary>
-    public class DbFixture : IDisposable
+    public class DbFixture : TestDbFixture
     {
-        public readonly string ConnectionString;
-        public readonly string CollabConnectionString;
-        public readonly NpgsqlFactory Factory;
-
-        public DbFixture()
-        {
-            ConnectionString = Environment.GetEnvironmentVariable("DatabaseConnectionString");
-            CollabConnectionString = Environment.GetEnvironmentVariable("CollaborationDatabaseConnectionString");
-            Factory = NpgsqlFactory.Instance;
-
-            Initialize(ConnectionString);
-            Initialize(CollabConnectionString);
-            ApplySchema();
-        }
-
-        /// <summary>
-        /// Ensure the database is able to receive connections before proceeding.
-        /// </summary>
-        public void Initialize(string connectionString)
-        {
-            var retries = 10;
-            var wait = 2000; // ms
-
-            while (retries >= 0)
-            {
-                try
-                {
-                    using (var conn = Factory.CreateConnection())
-                    {
-                        conn.ConnectionString = connectionString;
-                        conn.Open();
-                        conn.Close();
-
-                        return;
-                    }
-                }
-                catch (Npgsql.NpgsqlException ex)
-                {
-                    retries--;
-                    Console.WriteLine(ex.Message);
-                    Thread.Sleep(wait);
-                }
-            }
-        }
-
-        public void Dispose()
-        {
-            using (var conn = Factory.CreateConnection())
-            {
-                conn.ConnectionString = ConnectionString;
-                conn.Open();
-                conn.Execute("DROP TABLE IF EXISTS participants");
-                conn.Execute("DROP TABLE IF EXISTS uploads");
-                conn.Close();
-            }
-
-            using (var conn = Factory.CreateConnection())
-            {
-                conn.ConnectionString = CollabConnectionString;
-                conn.Open();
-                conn.Execute("DROP INDEX IF EXISTS index_match_id_on_match_res_events");
-                conn.Execute("DROP TABLE IF EXISTS match_res_events");
-                conn.Execute("DROP TABLE IF EXISTS matches");
-                conn.Close();
-            }
-
-        }
-
-        private void ApplySchema()
-        {
-            string perstateSql = System.IO.File.ReadAllText("per-state.sql", System.Text.Encoding.UTF8);
-            string matchesSql = System.IO.File.ReadAllText("match-record.sql", System.Text.Encoding.UTF8);
-
-            // Participants DB
-            using (var conn = Factory.CreateConnection())
-            {
-                conn.ConnectionString = ConnectionString;
-                conn.Open();
-
-                conn.Execute("DROP TABLE IF EXISTS participants");
-                conn.Execute("DROP TABLE IF EXISTS uploads");
-                conn.Execute(perstateSql);
-                conn.Execute("INSERT INTO uploads(created_at, publisher) VALUES(now() at time zone 'utc', current_user)");
-
-                conn.Close();
-            }
-
-            // Collaboration DB
-            using (var conn = Factory.CreateConnection())
-            {
-                conn.ConnectionString = CollabConnectionString;
-                conn.Open();
-
-                conn.Execute("DROP INDEX IF EXISTS index_match_id_on_match_res_events");
-                conn.Execute("DROP TABLE IF EXISTS match_res_events");
-                conn.Execute("DROP TABLE IF EXISTS matches");
-                conn.Execute(matchesSql);
-
-                conn.Close();
-            }
-        }
-
-        public void ClearParticipants()
-        {
-            using (var conn = Factory.CreateConnection())
-            {
-                conn.ConnectionString = ConnectionString;
-                conn.Open();
-
-                conn.Execute("DELETE FROM participants");
-
-                conn.Close();
-            }
-        }
-
-        public void ClearMatchRecords()
-        {
-            using (var conn = Factory.CreateConnection())
-            {
-                conn.ConnectionString = CollabConnectionString;
-                conn.Open();
-
-                conn.Execute("DELETE FROM matches");
-
-                conn.Close();
-            }
-        }
-
         public int CountMatchRecords()
         {
             int count;
